+++ src/test/groovy/ru/vyarus/java/generics/resolver/FailTests.groovy
@@ -3,6 +3,7 @@ package ru.vyarus.java.generics.resolver
 import ru.vyarus.java.generics.resolver.context.GenericsContext
 import ru.vyarus.java.generics.resolver.support.Base1
 import ru.vyarus.java.generics.resolver.support.BeanRoot
+import ru.vyarus.java.generics.resolver.support.Lvl2Base1
 import ru.vyarus.java.generics.resolver.support.Root
 import ru.vyarus.java.generics.resolver.support.noclash.NoClashRoot
 import ru.vyarus.java.generics.resolver.support.noclash.NoClashSub1
@@ -42,5 +43,14 @@ class FailTests extends Specification {
         then: "fail"
         thrown(UnknownGenericException)
 
+    }
+
+    def "Check type resolution fail"() {
+        when: "trying to access generic with bad name"
+        GenericsResolver.resolve(Root).resolveType(Lvl2Base1.getMethod("doSomth2").getGenericReturnType())
+        then: "fail"
+        def ex = thrown(UnknownGenericException)
+        ex.genericName == "I"
+        ex.contextType == Root
     }
 }